# frontend/types/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Swell Travel Agent API",
  "description": "Surf-forecast-driven travel discovery. All state-changing requests need the session cookie and an `X-CSRF-Token` header matching the `sta_csrf` cookie (GET `/api/auth/csrf` first).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/spots/{slug}/opportunities": {
   "get": {
    "tags": [
     "spots"
    ],
    "summary": "Spot Opportunities",
    "description": "The signed-in user's travel opportunities for this spot (empty when signed out).",
    "operationId": "spot_opportunities_api_spots__slug__opportunities_get",
    "parameters": [
     {
      "name": "slug",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Slug"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MatchSummary"
         },
         "title": "Response Spot Opportunities Api Spots  Slug  Opportunities Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/opportunities": {
   "get": {
    "tags": [
     "opportunities"
    ],
    "summary": "List Opportunities",
    "operationId": "list_opportunities_api_opportunities_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MatchStatus"
         }
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "include_past",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Include Past"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MatchSummary"
         },
         "title": "Response List Opportunities Api Opportunities Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/opportunities/{match_id}": {
   "get": {
    "tags": [
     "opportunities"
    ],
    "summary": "Get Opportunity",
    "operationId": "get_opportunity_api_opportunities__match_id__get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/opportunities/{match_id}/refresh-flights": {
   "post": {
    "tags": [
     "opportunities"
    ],
    "summary": "Refresh Flights",
    "operationId": "refresh_flights_api_opportunities__match_id__refresh_flights_post",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/opportunities/{match_id}/dismiss": {
   "post": {
    "tags": [
     "opportunities"
    ],
    "summary": "Dismiss",
    "operationId": "dismiss_api_opportunities__match_id__dismiss_post",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Match Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchSummary"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notifications": {
   "get": {
    "tags": [
     "notifications"
    ],
    "summary": "History",
    "operationId": "history_api_notifications_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/notifications/unsubscribe": {
   "post": {
    "tags": [
     "notifications"
    ],
    "summary": "Unsubscribe",
    "description": "Disable email alerts using the secret token from an alert email.\n\nSupports RFC 8058 one-click (form POST with ``List-Unsubscribe=One-Click`` and the\ntoken in the URL) and the web page (JSON body ``{\"token\": \"...\"}``). No session\nneeded: possession of the per-user token is the authorisation.",
    "operationId": "unsubscribe_api_notifications_unsubscribe_post",
    "parameters": [
     {
      "name": "token",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/dev/simulate-swell": {
   "post": {
    "tags": [
     "developer tools"
    ],
    "summary": "Simulate",
    "operationId": "simulate_api_dev_simulate_swell_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SimulateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Simulate Api Dev Simulate Swell Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/dev/outbox": {
   "get": {
    "tags": [
     "developer tools"
    ],
    "summary": "Outbox",
    "description": "Messages captured by the console email/SMS adapters (demo only).",
    "operationId": "outbox_api_dev_outbox_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Outbox Api Dev Outbox Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "DailySummary": {
    "properties": {
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "best_score": {
      "type": "integer",
      "title": "Best Score"
     },
     "best_label": {
      "type": "string",
      "title": "Best Label"
     },
     "best_time": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Best Time"
     },
     "height_min_ft": {
      "type": "number",
      "title": "Height Min Ft"
     },
     "height_max_ft": {
      "type": "number",
      "title": "Height Max Ft"
     },
     "avg_wind_kmh": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avg Wind Kmh"
     },
     "good_hours": {
      "type": "integer",
      "title": "Good Hours"
     },
     "confidence": {
      "type": "integer",
      "title": "Confidence"
     }
    },
    "type": "object",
    "required": [
     "date",
     "best_score",
     "best_label",
     "best_time",
     "height_min_ft",
     "height_max_ft",
     "avg_wind_kmh",
     "good_hours",
     "confidence"
    ],
    "title": "DailySummary"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LinkOut": {
    "properties": {
     "label": {
      "type": "string",
      "title": "Label"
     },
     "url": {
      "type": "string",
      "title": "Url"
     }
    },
    "type": "object",
    "required": [
     "label",
     "url"
    ],
    "title": "LinkOut"
   },
...
   "MatchDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "search_id": {
      "type": "string",
      "format": "uuid",
      "title": "Search Id"
     },
     "search_name": {
      "type": "string",
      "title": "Search Name"
     },
     "search_status": {
      "type": "string",
      "title": "Search Status"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "status_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status Reason"
     },
     "spot": {
      "$ref": "#/components/schemas/SpotMini"
     },
     "event_id": {
      "type": "string",
      "format": "uuid",
      "title": "Event Id"
     },
     "event_status": {
      "type": "string",
      "title": "Event Status"
     },
     "window_start": {
      "type": "string",
      "format": "date-time",
      "title": "Window Start"
     },
     "window_end": {
      "type": "string",
      "format": "date-time",
      "title": "Window End"
     },
     "qualifying_hours": {
      "type": "integer",
      "title": "Qualifying Hours"
     },
     "peak_score": {
      "type": "integer",
      "title": "Peak Score"
     },
     "avg_score": {
      "type": "number",
      "title": "Avg Score"
     },
     "peak_label": {
      "type": "string",
      "title": "Peak Label"
     },
     "breaking_height_min_ft": {
      "type": "number",
      "title": "Breaking Height Min Ft"
     },
     "breaking_height_max_ft": {
      "type": "number",
      "title": "Breaking Height Max Ft"
     },
     "confidence": {
      "type": "integer",
      "title": "Confidence"
     },
     "confidence_label": {
      "type": "string",
      "title": "Confidence Label"
     },
     "destination_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Destination Iata"
     },
     "transfer_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Transfer Minutes"
     },
     "origin_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Origin Iata"
     },
     "recommended_arrival_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Arrival Date"
     },
     "recommended_departure_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Departure Date"
     },
     "best_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Best Price"
     },
     "currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     },
     "overall_score": {
      "type": "integer",
      "title": "Overall Score"
     },
     "scores": {
      "additionalProperties": true,
      "type": "object",
      "title": "Scores"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "has_mock_flights": {
      "type": "boolean",
      "title": "Has Mock Flights"
     },
     "flight_checked_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Flight Checked At"
     },
     "notification_count": {
      "type": "integer",
      "title": "Notification Count"
     },
     "lead_days": {
      "type": "number",
      "title": "Lead Days"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "event": {
      "$ref": "#/components/schemas/EventOut"
     },
     "spot_airports": {
      "items": {
       "$ref": "#/components/schemas/SpotAirportOut"
      },
      "type": "array",
      "title": "Spot Airports"
     },
     "accessibility_notes": {
      "type": "string",
      "title": "Accessibility Notes"
     },
     "daily": {
      "items": {
       "$ref": "#/components/schemas/DailySummary"
      },
      "type": "array",
      "title": "Daily"
     },
     "offers": {
      "items": {
       "$ref": "#/components/schemas/OfferOut"
      },
      "type": "array",
      "title": "Offers"
     },
     "travel_window": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/TravelWindowOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "inspection_links": {
      "items": {
       "$ref": "#/components/schemas/LinkOut"
      },
      "type": "array",
      "title": "Inspection Links"
     },
     "limitations": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Limitations"
     },
     "flight_search_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Flight Search Error"
     }
    },
    "type": "object",
    "required": [
     "id",
     "search_id",
     "search_name",
     "search_status",
     "status",
     "status_reason",
     "spot",
     "event_id",
     "event_status",
     "window_start",
     "window_end",
     "qualifying_hours",
     "peak_score",
     "avg_score",
     "peak_label",
     "breaking_height_min_ft",
     "breaking_height_max_ft",
     "confidence",
     "confidence_label",
     "destination_iata",
     "transfer_minutes",
     "origin_iata",
     "recommended_arrival_date",
     "recommended_departure_date",
     "best_price",
     "currency",
     "overall_score",
     "scores",
     "is_demo",
     "has_mock_flights",
     "flight_checked_at",
     "notification_count",
     "lead_days",
     "created_at",
     "updated_at",
     "event",
     "spot_airports",
     "accessibility_notes",
     "daily",
     "offers",
     "travel_window",
     "inspection_links",
     "limitations",
     "flight_search_error"
    ],
    "title": "MatchDetail"
   },
   "MatchStatus": {
    "type": "string",
    "enum": [
     "pending_flights",
     "flight_found",
     "surf_only",
     "expired",
     "dismissed"
    ],
    "title": "MatchStatus"
   },
   "MatchSummary": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "search_id": {
      "type": "string",
      "format": "uuid",
      "title": "Search Id"
     },
     "search_name": {
      "type": "string",
      "title": "Search Name"
     },
     "search_status": {
      "type": "string",
      "title": "Search Status"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "status_reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status Reason"
     },
     "spot": {
      "$ref": "#/components/schemas/SpotMini"
     },
     "event_id": {
      "type": "string",
      "format": "uuid",
      "title": "Event Id"
     },
     "event_status": {
      "type": "string",
      "title": "Event Status"
     },
     "window_start": {
      "type": "string",
      "format": "date-time",
      "title": "Window Start"
     },
     "window_end": {
      "type": "string",
      "format": "date-time",
      "title": "Window End"
     },
     "qualifying_hours": {
      "type": "integer",
      "title": "Qualifying Hours"
     },
     "peak_score": {
      "type": "integer",
      "title": "Peak Score"
     },
     "avg_score": {
      "type": "number",
      "title": "Avg Score"
     },
     "peak_label": {
      "type": "string",
      "title": "Peak Label"
     },
     "breaking_height_min_ft": {
      "type": "number",
      "title": "Breaking Height Min Ft"
     },
     "breaking_height_max_ft": {
      "type": "number",
      "title": "Breaking Height Max Ft"
     },
     "confidence": {
      "type": "integer",
      "title": "Confidence"
     },
     "confidence_label": {
      "type": "string",
      "title": "Confidence Label"
     },
     "destination_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Destination Iata"
     },
     "transfer_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Transfer Minutes"
     },
     "origin_iata": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Origin Iata"
     },
     "recommended_arrival_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Arrival Date"
     },
     "recommended_departure_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Recommended Departure Date"
     },
     "best_price": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Best Price"
     },
     "currency": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency"
     },
     "overall_score": {
      "type": "integer",
      "title": "Overall Score"
     },
     "scores": {
      "additionalProperties": true,
      "type": "object",
      "title": "Scores"
     },
     "is_demo": {
      "type": "boolean",
      "title": "Is Demo"
     },
     "has_mock_flights": {
      "type": "boolean",
      "title": "Has Mock Flights"
     },
     "flight_checked_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Flight Checked At"
     },
     "notification_count": {
      "type": "integer",
      "title": "Notification Count"
     },
     "lead_days": {
      "type": "number",
      "title": "Lead Days"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "search_id",
     "search_name",
     "search_status",
     "status",
     "status_reason",
     "spot",
     "event_id",
     "event_status",
     "window_start",
     "window_end",
     "qualifying_hours",
     "peak_score",
     "avg_score",
     "peak_label",
     "breaking_height_min_ft",
     "breaking_height_max_ft",
     "confidence",
     "confidence_label",
     "destination_iata",
     "transfer_minutes",
     "origin_iata",
     "recommended_arrival_date",
     "recommended_departure_date",
     "best_price",
     "currency",
     "overall_score",
     "scores",
     "is_demo",
     "has_mock_flights",
     "flight_checked_at",
     "notification_count",
     "lead_days",
     "created_at",
     "updated_at"
    ],
    "title": "MatchSummary"
   },
   "Message": {
    "properties": {
     "message": {
      "type": "string",
      "title": "Message"
     }
    },
    "type": "object",
    "required": [
     "message"
    ],
    "title": "Message"
   },
...
   "NotificationPage": {
    "properties": {
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/NotificationOut"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "total",
     "items"
    ],
    "title": "NotificationPage"
   },
...
   "SimulateIn": {
    "properties": {
     "spot_slug": {
      "type": "string",
      "title": "Spot Slug"
     },
     "days_ahead": {
      "type": "number",
      "maximum": 14.0,
      "minimum": 1.0,
      "title": "Days Ahead",
      "default": 7.0
     },
     "duration_hours": {
      "type": "integer",
      "maximum": 168.0,
      "minimum": 12.0,
      "title": "Duration Hours",
      "default": 60
     },
     "target_breaking_ft": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 80.0,
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Target Breaking Ft"
     },
     "period_s": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 24.0,
        "minimum": 6.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Period S"
     },
     "run_pipeline": {
      "type": "boolean",
      "title": "Run Pipeline",
      "default": true
     },
     "deliver": {
      "type": "boolean",
      "title": "Deliver",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "spot_slug"
    ],
    "title": "SimulateIn"
   },
...
   "SpotAirportOut": {
    "properties": {
     "airport": {
      "$ref": "#/components/schemas/AirportOut"
     },
     "is_primary": {
      "type": "boolean",
      "title": "Is Primary"
     },
     "transfer_minutes": {
      "type": "integer",
      "title": "Transfer Minutes"
     },
     "transfer_mode": {
      "type": "string",
      "title": "Transfer Mode"
     },
     "notes": {
      "type": "string",
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "airport",
     "is_primary",
     "transfer_minutes",
     "transfer_mode",
     "notes"
    ],
    "title": "SpotAirportOut"
   },
...
   "SpotMini": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "slug": {
      "type": "string",
      "title": "Slug"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "country": {
      "type": "string",
      "title": "Country"
     },
     "country_code": {
      "type": "string",
      "title": "Country Code"
     },
     "region": {
      "type": "string",
      "title": "Region"
     },
     "region_group": {
      "type": "string",
      "title": "Region Group"
     },
     "latitude": {
      "type": "number",
      "title": "Latitude"
     },
     "longitude": {
      "type": "number",
      "title": "Longitude"
     },
     "timezone": {
      "type": "string",
      "title": "Timezone"
     },
     "break_type": {
      "type": "string",
      "title": "Break Type"
     }
    },
    "type": "object",
    "required": [
     "id",
     "slug",
     "name",
     "country",
     "country_code",
     "region",
     "region_group",
     "latitude",
     "longitude",
     "timezone",
     "break_type"
    ],
    "title": "SpotMini"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}